+++ walkeeper/src/bin/safekeeper.rs
@@ -10,7 +10,7 @@ use std::fs::File;
 use std::path::{Path, PathBuf};
 use std::thread;
 use tracing::*;
-use walkeeper::timeline::{CreateControlFile, FileStorage};
+use walkeeper::control_file::{self, CreateControlFile};
 use zenith_utils::http::endpoint;
 use zenith_utils::{logging, tcp_listener, GIT_VERSION};
 
@@ -96,7 +96,10 @@ fn main() -> Result<()> {
         .get_matches();
 
     if let Some(addr) = arg_matches.value_of("dump-control-file") {
-        let state = FileStorage::load_control_file(Path::new(addr), CreateControlFile::False)?;
+        let state = control_file::FileStorage::load_control_file(
+            Path::new(addr),
+            CreateControlFile::False,
+        )?;
         let json = serde_json::to_string(&state)?;
         print!("{}", json);
         return Ok(());
