+++ walkeeper/src/safekeeper.rs
@@ -3,7 +3,7 @@
 use anyhow::{bail, Context, Result};
 use byteorder::{LittleEndian, ReadBytesExt};
 use bytes::{Buf, BufMut, Bytes, BytesMut};
-use postgres_ffi::waldecoder::WalStreamDecoder;
+
 use postgres_ffi::xlog_utils::TimeLineID;
 use serde::{Deserialize, Serialize};
 use std::cmp::min;
@@ -13,12 +13,11 @@ use tracing::*;
 
 use lazy_static::lazy_static;
 
+use crate::control_file;
 use crate::send_wal::HotStandbyFeedback;
+use crate::wal_storage;
 use postgres_ffi::xlog_utils::MAX_SEND_SIZE;
-use zenith_metrics::{
-    register_gauge_vec, register_histogram_vec, Gauge, GaugeVec, Histogram, HistogramVec,
-    DISK_WRITE_SECONDS_BUCKETS,
-};
+use zenith_metrics::{register_gauge_vec, Gauge, GaugeVec};
 use zenith_utils::bin_ser::LeSer;
 use zenith_utils::lsn::Lsn;
 use zenith_utils::pq_proto::SystemId;
@@ -407,133 +406,87 @@ impl AcceptorProposerMessage {
     }
 }
 
-pub trait Storage {
-    /// Persist safekeeper state on disk.
-    fn persist(&mut self, s: &SafeKeeperState) -> Result<()>;
-    /// Write piece of wal in buf to disk and sync it.
-    fn write_wal(&mut self, server: &ServerInfo, startpos: Lsn, buf: &[u8]) -> Result<()>;
-    // Truncate WAL at specified LSN
-    fn truncate_wal(&mut self, s: &ServerInfo, endpos: Lsn) -> Result<()>;
-}
-
 lazy_static! {
     // The prometheus crate does not support u64 yet, i64 only (see `IntGauge`).
     // i64 is faster than f64, so update to u64 when available.
-    static ref FLUSH_LSN_GAUGE: GaugeVec = register_gauge_vec!(
-        "safekeeper_flush_lsn",
-        "Current flush_lsn, grouped by timeline",
-        &["tenant_id", "timeline_id"]
-    )
-    .expect("Failed to register safekeeper_flush_lsn gauge vec");
     static ref COMMIT_LSN_GAUGE: GaugeVec = register_gauge_vec!(
         "safekeeper_commit_lsn",
         "Current commit_lsn (not necessarily persisted to disk), grouped by timeline",
         &["tenant_id", "timeline_id"]
     )
     .expect("Failed to register safekeeper_commit_lsn gauge vec");
-    static ref WRITE_WAL_BYTES: HistogramVec = register_histogram_vec!(
-        "safekeeper_write_wal_bytes",
-        "Bytes written to WAL in a single request, grouped by timeline",
-        &["tenant_id", "timeline_id"],
-        vec![1.0, 10.0, 100.0, 1024.0, 8192.0, 128.0 * 1024.0, 1024.0 * 1024.0, 10.0 * 1024.0 * 1024.0]
-    )
-    .expect("Failed to register safekeeper_write_wal_bytes histogram vec");
-    static ref WRITE_WAL_SECONDS: HistogramVec = register_histogram_vec!(
-        "safekeeper_write_wal_seconds",
-        "Seconds spent writing and syncing WAL to a disk in a single request, grouped by timeline",
-        &["tenant_id", "timeline_id"],
-        DISK_WRITE_SECONDS_BUCKETS.to_vec()
-    )
-    .expect("Failed to register safekeeper_write_wal_seconds histogram vec");
 }
 
 struct SafeKeeperMetrics {
-    flush_lsn: Gauge,
     commit_lsn: Gauge,
-    write_wal_bytes: Histogram,
-    write_wal_seconds: Histogram,
 }
 
-struct SafeKeeperMetricsBuilder {
-    tenant_id: ZTenantId,
-    timeline_id: ZTimelineId,
-    flush_lsn: Lsn,
-    commit_lsn: Lsn,
-}
-
-impl SafeKeeperMetricsBuilder {
-    fn build(self) -> SafeKeeperMetrics {
-        let tenant_id = self.tenant_id.to_string();
-        let timeline_id = self.timeline_id.to_string();
-        let m = SafeKeeperMetrics {
-            flush_lsn: FLUSH_LSN_GAUGE.with_label_values(&[&tenant_id, &timeline_id]),
+impl SafeKeeperMetrics {
+    fn new(tenant_id: ZTenantId, timeline_id: ZTimelineId, commit_lsn: Lsn) -> Self {
+        let tenant_id = tenant_id.to_string();
+        let timeline_id = timeline_id.to_string();
+        let m = Self {
             commit_lsn: COMMIT_LSN_GAUGE.with_label_values(&[&tenant_id, &timeline_id]),
-            write_wal_bytes: WRITE_WAL_BYTES.with_label_values(&[&tenant_id, &timeline_id]),
-            write_wal_seconds: WRITE_WAL_SECONDS.with_label_values(&[&tenant_id, &timeline_id]),
         };
-        m.flush_lsn.set(u64::from(self.flush_lsn) as f64);
-        m.commit_lsn.set(u64::from(self.commit_lsn) as f64);
+        m.commit_lsn.set(u64::from(commit_lsn) as f64);
         m
     }
 }
 
 /// SafeKeeper which consumes events (messages from compute) and provides
 /// replies.
-pub struct SafeKeeper<ST: Storage> {
-    /// Locally flushed part of WAL with full records (end_lsn of last record).
-    /// Established by reading wal.
-    pub flush_lsn: Lsn,
+pub struct SafeKeeper<CTRL: control_file::Storage, WAL: wal_storage::Storage> {
     // Cached metrics so we don't have to recompute labels on each update.
     metrics: SafeKeeperMetrics,
+
     /// not-yet-flushed pairs of same named fields in s.*
     pub commit_lsn: Lsn,
     pub truncate_lsn: Lsn,
-    pub storage: ST,
     pub s: SafeKeeperState, // persistent part
-    decoder: WalStreamDecoder,
+
+    pub control_store: CTRL,
+    pub wal_store: WAL,
 }
 
-impl<ST> SafeKeeper<ST>
+impl<CTRL, WAL> SafeKeeper<CTRL, WAL>
 where
-    ST: Storage,
+    CTRL: control_file::Storage,
+    WAL: wal_storage::Storage,
 {
     // constructor
     pub fn new(
         ztli: ZTimelineId,
-        flush_lsn: Lsn,
-        storage: ST,
+        control_store: CTRL,
+        wal_store: WAL,
         state: SafeKeeperState,
-    ) -> SafeKeeper<ST> {
+    ) -> SafeKeeper<CTRL, WAL> {
         if state.server.timeline_id != ZTimelineId::from([0u8; 16])
             && ztli != state.server.timeline_id
         {
             panic!("Calling SafeKeeper::new with inconsistent ztli ({}) and SafeKeeperState.server.timeline_id ({})", ztli, state.server.timeline_id);
         }
+
         SafeKeeper {
-            flush_lsn,
-            metrics: SafeKeeperMetricsBuilder {
-                tenant_id: state.server.tenant_id,
-                timeline_id: ztli,
-                flush_lsn,
-                commit_lsn: state.commit_lsn,
-            }
-            .build(),
+            metrics: SafeKeeperMetrics::new(state.server.tenant_id, ztli, state.commit_lsn),
             commit_lsn: state.commit_lsn,
             truncate_lsn: state.truncate_lsn,
-            storage,
             s: state,
-            decoder: WalStreamDecoder::new(Lsn(0)),
+            control_store,
+            wal_store,
         }
     }
 
     /// Get history of term switches for the available WAL
     fn get_term_history(&self) -> TermHistory {
-        self.s.acceptor_state.term_history.up_to(self.flush_lsn)
+        self.s
+            .acceptor_state
+            .term_history
+            .up_to(self.wal_store.flush_lsn())
     }
 
     #[cfg(test)]
     fn get_epoch(&self) -> Term {
-        self.s.acceptor_state.get_epoch(self.flush_lsn)
+        self.s.acceptor_state.get_epoch(self.wal_store.flush_lsn())
     }
 
     /// Process message from proposer and possibly form reply. Concurrent
@@ -575,21 +528,20 @@ where
         }
 
         // set basic info about server, if not yet
+        // TODO: verify that is doesn't change after
         self.s.server.system_id = msg.system_id;
         self.s.server.tenant_id = msg.tenant_id;
         self.s.server.timeline_id = msg.ztli;
         self.s.server.wal_seg_size = msg.wal_seg_size;
-        self.storage
+        self.control_store
             .persist(&self.s)
             .context("failed to persist shared state")?;
 
-        self.metrics = SafeKeeperMetricsBuilder {
-            tenant_id: self.s.server.tenant_id,
-            timeline_id: self.s.server.timeline_id,
-            flush_lsn: self.flush_lsn,
-            commit_lsn: self.commit_lsn,
-        }
-        .build();
+        // pass wal_seg_size to read WAL and find flush_lsn
+        self.wal_store.init_storage(&self.s)?;
+
+        // update tenant_id/timeline_id in metrics
+        self.metrics = SafeKeeperMetrics::new(msg.tenant_id, msg.ztli, self.commit_lsn);
 
         info!(
             "processed greeting from proposer {:?}, sending term {:?}",
@@ -609,14 +561,14 @@ where
         let mut resp = VoteResponse {
             term: self.s.acceptor_state.term,
             vote_given: false as u64,
-            flush_lsn: self.flush_lsn,
+            flush_lsn: self.wal_store.flush_lsn(),
             truncate_lsn: self.s.truncate_lsn,
             term_history: self.get_term_history(),
         };
         if self.s.acceptor_state.term < msg.term {
             self.s.acceptor_state.term = msg.term;
             // persist vote before sending it out
-            self.storage.persist(&self.s)?;
+            self.control_store.persist(&self.s)?;
             resp.term = self.s.acceptor_state.term;
             resp.vote_given = true as u64;
         }
@@ -628,7 +580,7 @@ where
     fn bump_if_higher(&mut self, term: Term) -> Result<()> {
         if self.s.acceptor_state.term < term {
             self.s.acceptor_state.term = term;
-            self.storage.persist(&self.s)?;
+            self.control_store.persist(&self.s)?;
         }
         Ok(())
     }
@@ -637,7 +589,7 @@ where
     fn append_response(&self) -> AppendResponse {
         AppendResponse {
             term: self.s.acceptor_state.term,
-            flush_lsn: self.flush_lsn,
+            flush_lsn: self.wal_store.flush_lsn(),
             commit_lsn: self.s.commit_lsn,
             // will be filled by the upper code to avoid bothering safekeeper
             hs_feedback: HotStandbyFeedback::empty(),
@@ -653,22 +605,12 @@ where
             return Ok(None);
         }
 
-        // TODO: cross check divergence point
+        // truncate wal, update the lsns
+        self.wal_store.truncate_wal(msg.start_streaming_at)?;
 
-        // streaming must not create a hole
-        assert!(self.flush_lsn == Lsn(0) || self.flush_lsn >= msg.start_streaming_at);
-
-        // truncate obsolete part of WAL
-        if self.flush_lsn != Lsn(0) {
-            self.storage
-                .truncate_wal(&self.s.server, msg.start_streaming_at)?;
-        }
-        // update our end of WAL pointer
-        self.flush_lsn = msg.start_streaming_at;
-        self.metrics.flush_lsn.set(u64::from(self.flush_lsn) as f64);
         // and now adopt term history from proposer
         self.s.acceptor_state.term_history = msg.term_history.clone();
-        self.storage.persist(&self.s)?;
+        self.control_store.persist(&self.s)?;
 
         info!("start receiving WAL since {:?}", msg.start_streaming_at);
 
@@ -694,42 +636,14 @@ where
         // After ProposerElected, which performs truncation, we should get only
         // indeed append requests (but flush_lsn is advanced only on record
         // boundary, so might be less).
-        assert!(self.flush_lsn <= msg.h.begin_lsn);
+        assert!(self.wal_store.flush_lsn() <= msg.h.begin_lsn);
 
         self.s.proposer_uuid = msg.h.proposer_uuid;
         let mut sync_control_file = false;
 
         // do the job
-        let mut last_rec_lsn = Lsn(0);
         if !msg.wal_data.is_empty() {
-            self.metrics
-                .write_wal_bytes
-                .observe(msg.wal_data.len() as f64);
-            {
-                let _timer = self.metrics.write_wal_seconds.start_timer();
-                self.storage
-                    .write_wal(&self.s.server, msg.h.begin_lsn, &msg.wal_data)?;
-            }
-
-            // figure out last record's end lsn for reporting (if we got the
-            // whole record)
-            if self.decoder.available() != msg.h.begin_lsn {
-                info!(
-                    "restart decoder from {} to {}",
-                    self.decoder.available(),
-                    msg.h.begin_lsn,
-                );
-                self.decoder = WalStreamDecoder::new(msg.h.begin_lsn);
-            }
-            self.decoder.feed_bytes(&msg.wal_data);
-            loop {
-                match self.decoder.poll_decode()? {
-                    None => break, // no full record yet
-                    Some((lsn, _rec)) => {
-                        last_rec_lsn = lsn;
-                    }
-                }
-            }
+            self.wal_store.write_wal(msg.h.begin_lsn, &msg.wal_data)?;
 
             // If this was the first record we ever receieved, remember LSN to help
             // find_end_of_wal skip the hole in the beginning.
@@ -739,16 +653,11 @@ where
             }
         }
 
-        if last_rec_lsn > self.flush_lsn {
-            self.flush_lsn = last_rec_lsn;
-            self.metrics.flush_lsn.set(u64::from(self.flush_lsn) as f64);
-        }
-
         // Advance commit_lsn taking into account what we have locally.
         // commit_lsn can be 0, being unknown to new walproposer while he hasn't
         // collected majority of its epoch acks yet, ignore it in this case.
         if msg.h.commit_lsn != Lsn(0) {
-            let commit_lsn = min(msg.h.commit_lsn, self.flush_lsn);
+            let commit_lsn = min(msg.h.commit_lsn, self.wal_store.flush_lsn());
             // If new commit_lsn reached epoch switch, force sync of control
             // file: walproposer in sync mode is very interested when this
             // happens. Note: this is for sync-safekeepers mode only, as
@@ -774,7 +683,7 @@ where
         }
 
         if sync_control_file {
-            self.storage.persist(&self.s)?;
+            self.control_store.persist(&self.s)?;
         }
 
         let resp = self.append_response();
@@ -793,34 +702,52 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::wal_storage::Storage;
 
     // fake storage for tests
-    struct InMemoryStorage {
+    struct InMemoryState {
         persisted_state: SafeKeeperState,
     }
 
-    impl Storage for InMemoryStorage {
+    impl control_file::Storage for InMemoryState {
         fn persist(&mut self, s: &SafeKeeperState) -> Result<()> {
             self.persisted_state = s.clone();
             Ok(())
         }
+    }
 
-        fn write_wal(&mut self, _server: &ServerInfo, _startpos: Lsn, _buf: &[u8]) -> Result<()> {
+    struct DummyWalStore {
+        lsn: Lsn,
+    }
+
+    impl wal_storage::Storage for DummyWalStore {
+        fn flush_lsn(&self) -> Lsn {
+            self.lsn
+        }
+
+        fn init_storage(&mut self, _state: &SafeKeeperState) -> Result<()> {
             Ok(())
         }
 
-        fn truncate_wal(&mut self, _server: &ServerInfo, _end_pos: Lsn) -> Result<()> {
+        fn write_wal(&mut self, startpos: Lsn, buf: &[u8]) -> Result<()> {
+            self.lsn = startpos + buf.len() as u64;
+            Ok(())
+        }
+
+        fn truncate_wal(&mut self, end_pos: Lsn) -> Result<()> {
+            self.lsn = end_pos;
             Ok(())
         }
     }
 
     #[test]
     fn test_voting() {
-        let storage = InMemoryStorage {
+        let storage = InMemoryState {
             persisted_state: SafeKeeperState::new(),
         };
+        let wal_store = DummyWalStore { lsn: Lsn(0) };
         let ztli = ZTimelineId::from([0u8; 16]);
-        let mut sk = SafeKeeper::new(ztli, Lsn(0), storage, SafeKeeperState::new());
+        let mut sk = SafeKeeper::new(ztli, storage, wal_store, SafeKeeperState::new());
 
         // check voting for 1 is ok
         let vote_request = ProposerAcceptorMessage::VoteRequest(VoteRequest { term: 1 });
@@ -831,11 +758,11 @@ mod tests {
         }
 
         // reboot...
-        let state = sk.storage.persisted_state.clone();
-        let storage = InMemoryStorage {
+        let state = sk.control_store.persisted_state.clone();
+        let storage = InMemoryState {
             persisted_state: state.clone(),
         };
-        sk = SafeKeeper::new(ztli, Lsn(0), storage, state);
+        sk = SafeKeeper::new(ztli, storage, sk.wal_store, state);
 
         // and ensure voting second time for 1 is not ok
         vote_resp = sk.process_msg(&vote_request);
@@ -847,11 +774,12 @@ mod tests {
 
     #[test]
     fn test_epoch_switch() {
-        let storage = InMemoryStorage {
+        let storage = InMemoryState {
             persisted_state: SafeKeeperState::new(),
         };
+        let wal_store = DummyWalStore { lsn: Lsn(0) };
         let ztli = ZTimelineId::from([0u8; 16]);
-        let mut sk = SafeKeeper::new(ztli, Lsn(0), storage, SafeKeeperState::new());
+        let mut sk = SafeKeeper::new(ztli, storage, wal_store, SafeKeeperState::new());
 
         let mut ar_hdr = AppendRequestHeader {
             term: 1,
@@ -892,7 +820,7 @@ mod tests {
         };
         let resp = sk.process_msg(&ProposerAcceptorMessage::AppendRequest(append_request));
         assert!(resp.is_ok());
-        sk.flush_lsn = Lsn(3); // imitate the complete record at 3 %)
+        sk.wal_store.truncate_wal(Lsn(3)).unwrap(); // imitate the complete record at 3 %)
         assert_eq!(sk.get_epoch(), 1);
     }
 }
